perf: 优化文件下载

+++ weiboSpider.py
@@ -400,11 +400,12 @@ class Weibo(object):
     def download_one_file(self, url, file_path, type, weibo_id):
         """下载单个文件(图片/视频)"""
         try:
-            s = requests.Session()
-            s.mount(url, HTTPAdapter(max_retries=5))
-            downloaded = s.get(url, timeout=(5, 10))
-            with open(file_path, 'wb') as f:
-                f.write(downloaded.content)
+            if not os.path.isfile(file_path):
+                s = requests.Session()
+                s.mount(url, HTTPAdapter(max_retries=5))
+                downloaded = s.get(url, timeout=(5, 10))
+                with open(file_path, 'wb') as f:
+                    f.write(downloaded.content)
         except Exception as e:
             error_file = self.get_filepath(
                 type) + os.sep + 'not_downloaded.txt'
